+++ tooling/postcss-tailwind-css-modules/tsup.config.ts
@@ -10,22 +10,21 @@
  * governing permissions and limitations under the License.
  */
 
-import baseConfig from '@accelint/vitest-config/dom';
-import tailwindcss from '@tailwindcss/vite';
-import { defineConfig, mergeConfig } from 'vitest/config';
+import { defineConfig } from 'tsup';
 
-export default mergeConfig(
-  baseConfig,
-  defineConfig({
-    plugins: [tailwindcss()],
-    test: {
-      globals: true,
-      setupFiles: './src/test/setup.ts',
-      environment: 'jsdom',
-      exclude: ['./storybook-static'],
-      typecheck: {
-        tsconfig: './tsconfig.dev.json',
-      },
-    },
-  }),
-);
+export default defineConfig([
+  {
+    format: ['esm'],
+    minify: true,
+    dts: true,
+    external: ['postcss-selector-parser'],
+    entry: ['src/index.ts'],
+  },
+  {
+    format: ['cjs'],
+    minify: true,
+    dts: true,
+    external: ['postcss-selector-parser'],
+    entry: ['src/index.cts'],
+  },
+]);